Commented code out.
Fixed the data download/reading .csv part of the script.

Set the script up to make a forecast from 'today' onward. It does so, but only for five days for whatever reason. This appears to be the big issue moving forward.

I did not work on getting the old forecast data to make up for the gap in data between the forecast today, and the end of the NEON data. I don't think that period of time should be a problem as the historical fit data should tune to parameters just fine for the forecast moving forward from 'today'.